% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{dataDanube}
\alias{dataDanube}
\title{Danube river discharges}
\source{
Bavarian Environmental Agency (\url{http://www.gkd.bayern.de})
}
\usage{
graphicalExtremes::danube
}
\description{
Daily river discharges, measured in \eqn{m^3/s}, at 31 stations spread over the upper Danube basin. The data set covers the period from 1960 to 2010 but only the months of June, July, and August are retained. These data are already declustered following  Mhalla et al.(2020) methodology.
}
\examples{
graphicalExtremes::danube
}
\references{
Asadi, P., Davison, A.C.,  and Engelke, S. (2015). Extremes on river networks. The Annals of Applied Statistics, 9(4), 2023-2050.
}
