% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{FrechetMargin}
\alias{FrechetMargin}
\title{Unit-Frechet transformation}
\usage{
FrechetMargin(X)
}
\arguments{
\item{X}{a numeric vector}
}
\description{
Transforms data to unit-Frechet scale using rank transformation
}
\examples{
FrechetMargin()
}
\references{
Legrand et al.
}
